+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update validator cloud-init to support 20.18 release
 - In restore task, add option to change the software version of control components and SD-WAN/SD-Routing edges during restore
 - Increase SD-WAN Manager CPU count from 8 to 10
+- In setup and restore tasks, and check to make sure user is not using default credentials for SD-WAN Manager
 
 # Catalyst SD-WAN Lab 2.0.15 [Feb 25, 2025]